requirements.txt
flask==2.0.1
werkzeug==2.0.3
web3
pyqrcode==1.2.1
pypng==0.0.21
py-solc-x==1.1.1
Pillow==9.0.1
streamlit==1.22.0
requests==2.28.1
numpy
opencv-python
pymysql
toml==0.10.2
parsimonious>=0.8.0,<0.9.0
streamlit_webrtc